+++ packages/config/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @opengeni/config
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [231b103]
+- Updated dependencies [392c575]
+- Updated dependencies [9827c25]
+- Updated dependencies [5904fd1]
+- Updated dependencies [14dd6fe]
+  - @opengeni/contracts@2.15.0
+
 ## 1.0.1
 
 ### Patch Changes